refactor(desktop): the brain's app acts have a carrier
